refactor: implementar composable para armazenar erros e resultados da execução de tasks em variáveis de estado
